Fix #1784: Refactor mentorship app admin.py into separate files (#1952)
* Refactor mentorship admin.py

* Update code

---------
