+++ backend/apps/mentorship/admin/__init__.py
@@ -1,0 +1,8 @@
+"""Mentorship app admin."""
+
+from .mentee import MenteeAdmin
+from .mentee_program import MenteeProgramAdmin
+from .mentor import MentorAdmin
+from .module import ModuleAdmin
+from .program import ProgramAdmin
+from .task import TaskAdmin
